New upstream version 5.7.0

+++ src/modules/nat_traversal/nat_traversal.c
@@ -39,9 +39,11 @@
 #include <stdlib.h>
 #include <unistd.h>
 #include <string.h>
+#include <stdint.h>
 #include <ctype.h>
 #include <errno.h>
 #include <time.h>
+#include <inttypes.h>
 #include <sys/types.h>
 #include <arpa/inet.h>
 
@@ -922,7 +924,7 @@ static time_t get_register_expire(
 		}
 	}
 
-	LM_DBG("maximum expire for all contacts: %u\n", (unsigned)expire);
+	LM_DBG("maximum expire for all contacts: %" PRIu64 "\n", (uint64_t)expire);
 
 	return (expire ? expire + now : 0);
 }
